Add support for SQL Server Vector Search (#722)
## Motivation and Context (Why the change? What's the scenario?)

Azure SQL Database now provides native Vector Support, currently in EAP
(https://devblogs.microsoft.com/azure-sql/announcing-eap-native-vector-support-in-azure-sql-database).

This PR extends SQL Server Memory DB with a flag that allows to use this
new feature.

> [!IMPORTANT]  
> Remember that, at this time, Vector Support is available only on Azure
SQL Database. On the other hand, the current SQL Server Memory DB
requires a COLUMNSTORE INDEX that, on Azure, is available only on vCore
databases and Standard databases in S3 and above pricing tiers
(https://azure.microsoft.com/en-us/blog/columnstore-support-in-standard-tier-azure-sql-databases).

---------
