+++ service/Service/appsettings.json
@@ -577,7 +577,9 @@
         "MemoryCollectionTableName": "KMCollections",
         "MemoryTableName": "KMMemories",
         "EmbeddingsTableName": "KMEmbeddings",
-        "TagsTableName": "KMMemoriesTags"
+        "TagsTableName": "KMMemoriesTags",
+        // See https://devblogs.microsoft.com/azure-sql/announcing-eap-native-vector-support-in-azure-sql-database
+        "UseNativeVectorSearch": false
       }
     }
   }
